pyproject.toml
[project]
name = "ncc-school-management"
version = "0.1.0"
description = "NCC School Management System"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "psycopg2-binary>=2.9.0",
    "django-cors-headers>=4.3.0",
    "django-extensions>=3.2.0",
    "django-filter>=25.1",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.0.0",
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ncc_school_management"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
    ".venv",
    "node_modules",
]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ncc_school_management.test_settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short --strict-markers --disable-warnings --nomigrations --reuse-db"
testpaths = [".", "api", "comercial", "common", "crm", "financial", "management"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]
